demo 2

+++ src/main/java/com/hanwen/controllers/PlanController.java
@@ -9,6 +9,7 @@ import com.hanwen.model.Plan;
 import com.hanwen.repositories.PlanRepository;
 import com.hanwen.service.PlanService;
 
+import org.apache.commons.codec.digest.DigestUtils;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -16,6 +17,7 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.context.request.WebRequest;
 
 
 @RestController
@@ -43,7 +45,7 @@ public class PlanController {
     }
 
     @RequestMapping(value = "/plan/{id}", method = RequestMethod.PUT)
-    public ResponseEntity update(@PathVariable("id") String id, @RequestBody String json) {
+    public ResponseEntity update(@PathVariable("id") String id, @RequestBody String json, WebRequest request) {
         Plan plan = planRepository.findById(id);
         if(plan == null) {
             return new ResponseEntity<>(null, HttpStatus.NO_CONTENT);
@@ -53,10 +55,19 @@ public class PlanController {
         try {
             Plan data = gson.fromJson(json, Plan.class);
 
+            if (request.checkNotModified(plan.getObjectId())) {
+                // 2. shortcut exit - no further processing necessary
+                //  it will also convert the response to an 304 Not Modified
+                //  with an empty body
+                return new ResponseEntity<>(plan, HttpStatus.NOT_MODIFIED);
+            }
+
             planRepository.delete(id);
             planRepository.save(data);
 
-            return new ResponseEntity<>(planRepository.findById(id), HttpStatus.OK);
+            return ResponseEntity.ok()
+            .eTag(data.getObjectId())
+            .body(data);
         } catch(Exception e) {
             return new ResponseEntity<>("Error when parsing JSON", HttpStatus.BAD_REQUEST);
         }
@@ -68,19 +79,27 @@ public class PlanController {
         if(map.isEmpty()) {
             return new ResponseEntity<>(null, HttpStatus.NO_CONTENT);
         }
+
         return ResponseEntity.ok()
         .eTag(String.valueOf(map))
         .body(map);
     }
 
     @RequestMapping(value = "/plan/{id}", method = RequestMethod.GET)
-    public ResponseEntity<Plan> findById(@PathVariable("id") String id) {
+    public ResponseEntity<Plan> findById(@PathVariable("id") String id, WebRequest request) {
         Plan plan = planRepository.findById(id);
         if(plan == null) {
             return new ResponseEntity<>(null, HttpStatus.NO_CONTENT);
         }
+        if (request.checkNotModified(plan.getObjectId())) {
+            // 2. shortcut exit - no further processing necessary
+            //  it will also convert the response to an 304 Not Modified
+            //  with an empty body
+            return new ResponseEntity<>(plan, HttpStatus.NOT_MODIFIED);
+        }
+
         return ResponseEntity.ok()
-        .eTag(String.valueOf(plan))
+        .eTag(plan.getObjectId())
         .body(plan);
     }
 
@@ -95,7 +114,7 @@ public class PlanController {
     }
 
     @RequestMapping(value = "/plan/{id}", method = RequestMethod.PATCH)
-    public ResponseEntity patch(@PathVariable("id") String id, @RequestBody String json) {
+    public ResponseEntity patch(@PathVariable("id") String id, @RequestBody String json, WebRequest request) {
         Plan plan = planRepository.findById(id);
         if(plan == null) {
             return new ResponseEntity<>(null, HttpStatus.NO_CONTENT);
@@ -103,24 +122,29 @@ public class PlanController {
 
         Gson gson = new Gson();
         try {
-            LinkedPlanServices[] data = gson.fromJson(json, LinkedPlanServices[].class);
+            Plan data = gson.fromJson(json, Plan.class);
 
             List<LinkedPlanServices> list = plan.getLinkedPlanServices();
 
-            for(LinkedPlanServices lps : data) {
+            for(LinkedPlanServices lps : data.getLinkedPlanServices()) {
                 list.add(lps);
+            }
+
+            if (request.checkNotModified(plan.getObjectId())) {
+                // 2. shortcut exit - no further processing necessary
+                //  it will also convert the response to an 304 Not Modified
+                //  with an empty body
+                return new ResponseEntity<>(plan, HttpStatus.NOT_MODIFIED);
             }
 
             planRepository.save(plan);
 
-            return new ResponseEntity<>(planRepository.findById(id), HttpStatus.OK);
+            return ResponseEntity.ok()
+            .eTag(plan.getObjectId())
+            .body(plan);
+            // return new ResponseEntity<>(planRepository.findById(id), HttpStatus.OK);
         } catch(Exception e) {
             return new ResponseEntity<>("Error when parsing JSON", HttpStatus.BAD_REQUEST);
         }
     }
-
-    // @RequestMapping(value = "/registry", method = RequestMethod.GET)
-    // public RedirectView redirect() {
-    //     return new RedirectView("/plan");
-    // }
 }
